chore(aspenetcore): update aspnetcore projects for 1.0
