--- skeleton-esnext-aspnetcore/src/skeleton/bower.json
@@ -1,10 +1,0 @@
-{
-  "name": "ASP.NET",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "3.3.5",
-    "jquery": "2.1.4",
-    "jquery-validation": "1.14.0",
-    "jquery-validation-unobtrusive": "3.2.4"
-  }
-}
