+++ skeleton-typescript-aspnetcore/src/skeleton/gulpfile.js
@@ -1,4 +1,4 @@
-﻿/// <binding BeforeBuild='build' Clean='clean' />
+/// <binding BeforeBuild='build' Clean='clean' ProjectOpened='watch' />
 // all gulp tasks are located in the ./build/tasks directory
 // gulp configuration is in files in ./build directory
 require('require-dir')('build/tasks');
