+++ skeleton-typescript-aspnetcore/src/skeleton/package.json
@@ -12,16 +12,13 @@
     "gulp": "^3.9.1",
     "gulp-bump": "^2.1.0",
     "gulp-changed": "^1.3.0",
-    "gulp-concat": "2.6.0",
-    "gulp-cssmin": "0.1.7",
     "gulp-htmlmin": "^2.0.0",
     "gulp-notify": "^2.2.0",
     "gulp-plumber": "^1.1.0",
-    "gulp-protractor": "^2.4.0",
+    "gulp-protractor": "2.4.0",
     "gulp-sourcemaps": "^1.6.0",
-    "gulp-typescript": "^2.13.6",
     "gulp-tslint": "^5.0.0",
-    "gulp-uglify": "1.5.3",
+    "gulp-typescript": "^2.13.6",
     "isparta": "^4.0.0",
     "jasmine-core": "^2.4.1",
     "jspm": "^0.16.15",
@@ -31,11 +28,11 @@
     "karma-systemjs": "^0.14.0",
     "object.assign": "^4.0.3",
     "require-dir": "^0.3.0",
-    "rimraf": "2.5.2",
     "run-sequence": "^1.2.1",
     "systemjs": "0.19.31",
     "tslint": "^3.11.0",
     "typescript": ">=1.9.0-dev || ^2.0.0",
+    "typings": "^1.3.0",
     "vinyl-paths": "^2.1.0",
     "yargs": "^4.7.1"
   },
@@ -47,7 +44,9 @@
     "vscode"
   ],
   "scripts": {
-    "postinstall": "./node_modules/.bin/typings install",
+    "postinstall": "npm run jspm-install && npm run typings-install",
+    "jspm-install": "./node_modules/.bin/jspm install -y",
+    "typings-install": "./node_modules/.bin/typings install",
     "test": "gulp test",
     "e2e": "gulp serve e2e"
   },
