+++ skeleton-typescript-aspnetcore/src/skeleton/project.json
@@ -105,7 +105,7 @@
     },
 
     "scripts": {
-        "prepublish": [ "npm install", "bower install", "gulp clean", "gulp min" ],
+        "prepublish": [ "npm install", "gulp build"],
         "postpublish": [ "dotnet publish-iis --publish-folder %publish:OutputPath% --framework %publish:FullTargetFramework%" ]
     }
 }
