+++ skeleton-typescript-aspnetcore/src/skeleton/src/welcome.ts
@@ -1,7 +1,7 @@
 //import {computedFrom} from 'aurelia-framework';
 
 export class Welcome {
-  heading = 'Welcome to the Aurelia Navigation App [skeleton-typescript-aspnetcore]!';
+  heading = 'Welcome to the Aurelia Navigation App!';
   firstName = 'John';
   lastName = 'Doe';
   previousValue = this.fullName;
