+++ skeleton-typescript-aspnetcore/src/skeleton/tsconfig.json
@@ -1,30 +1,29 @@
 {
-  "version": "1.7.3",
   "compileOnSave": false,
   "compilerOptions": {
     "rootDir": "src/",
+    "outDir": "wwwroot/dist",
     "sourceMap": true,
     "target": "es5",
     "module": "amd",
     "declaration": false,
     "noImplicitAny": false,
-    "noResolve": true,
     "removeComments": true,
-    "noLib": false,
     "emitDecoratorMetadata": true,
-    "experimentalDecorators": true
+    "experimentalDecorators": true,
+    "moduleResolution": "node",
+    "lib": ["es2015", "dom"]
   },
   "filesGlob": [
     "./src/**/*.ts",
     "./test/**/*.ts",
-    "./typings/browser.d.ts",
+    "./typings/index.d.ts",
     "./custom_typings/**/*.d.ts",
     "./wwwroot/jspm_packages/**/*.d.ts"
   ],
   "exclude": [
     "node_modules",
-    "typings/main",
-    "typings/main.d.ts"
+    "wwwroot/jspm_packages"
   ],
   "atom": {
     "rewriteTsconfig": false
